[fix] provider issues and added Add button to EntityTypes
