+++ CHANGELOG.md
@@ -1,2 +1,5 @@
+## Additions 🍎
+- PlayerEX is now on 1.21.1 for fabric & neoforge!
+- Add button for Entity Types (will downstream this to 1.20.1 momentarily).
 ## Changes 🌽
-- Added extension (reified) functions that could prove to be useful to Kotlin users.
+This update affected a sector of mixins which I have redone, but I will keep it in beta until its stability is ensured along with PlayerEX.
